tournament_service/requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==6.0
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==25.0
pathlib==1.0.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
setuptools==80.9.0
sqlparse==0.5.4
wheel==0.45.1
pytest-order==1.3.0
drf-spectacular==0.27.2

